Make module structure for loading
