+++ vegans/utils/loading/__init__.py
@@ -1,0 +1,6 @@
+from vegans.utils.loading.ExampleLoader import ExampleLoader
+from vegans.utils.loading.MNISTLoader import MNISTLoader
+from vegans.utils.loading.FashionMNISTLoader import FashionMNISTLoader
+from vegans.utils.loading.CIFAR10Loader import CIFAR10Loader
+from vegans.utils.loading.CIFAR100Loader import CIFAR100Loader
+from vegans.utils.loading.CelebALoader import CelebALoader
