+++ vegans/utils/loading/architectures/example.py
