+++ vegans/utils/loading/architectures/mnist.py
